Export computeTupleSubsetHash
